Some work on Main.py
