+++ NdbClasses.py
@@ -1,0 +1,36 @@
+from google.appengine.ext import ndb
+
+
+class Stream(ndb.Model):
+    owner = ndb.KeyProperty(indexed=True, kind='StreamUser')
+    name = ndb.StringProperty(indexed=False)
+    coverImageURL = ndb.StringProperty(indexed=False)
+    numViews = ndb.IntegerProperty(indexed=False)
+
+
+class StreamItem(ndb.Model):
+    stream = ndb.KeyProperty(indexed=True, kind='Stream')
+    name = ndb.StringProperty(indexed=False)
+    image = ndb.BlobProperty(indexed=False)
+    dateAdded = ndb.DateProperty(indexed=False)
+
+
+class Tag(ndb.Model):
+    name = ndb.StringProperty(indexed=False)
+
+
+class StreamTag(ndb.Model):
+    stream = ndb.KeyProperty(indexed=True, kind='Stream')
+    tag = ndb.KeyProperty(indexed=True, kind='Tag')
+
+
+class StreamSubscriber(ndb.Model):
+    stream = ndb.KeyProperty(indexed=True, kind='Stream')
+    user = ndb.KeyProperty(indexed=True, kind='StreamUser')
+
+
+class StreamUser(ndb.Model):
+    firstName = ndb.StringProperty(indexed=False)
+    lastName = ndb.StringProperty(indexed=False)
+    nickName = ndb.StringProperty(indexed=False)
+    email = ndb.StringProperty(indexed=False)
